Make image cache per-document rather than global
