+++ components/net_traits/Cargo.toml
@@ -16,6 +16,7 @@ heapsize_derive = "0.1"
 hyper = "0.9.9"
 hyper_serde = "0.5"
 image = "0.12"
+immeta = "0.3.1"
 ipc-channel = "0.7"
 lazy_static = "0.2"
 log = "0.3.5"
